add datasource and UI for AQI

+++ lib/feature/weather/data/model/weather_forecast_model.dart
@@ -82,20 +82,20 @@ class DayModel extends DayEntity {
   final double? mintempF;
   final double? avgtempC;
   final double? avgtempF;
-  final int? maxwindMph;
+  final double? maxwindMph;
   final double? maxwindKph;
   final double? totalprecipMm;
   final double? totalprecipIn;
-  final int? totalsnowCm;
-  final int? avgvisKm;
-  final int? avgvisMiles;
+  final double? totalsnowCm;
+  final double? avgvisKm;
+  final double? avgvisMiles;
   final int? avghumidity;
   final int? dailyWillItRain;
   final int? dailyChanceOfRain;
   final int? dailyWillItSnow;
   final int? dailyChanceOfSnow;
   final ConditionModel? condition;
-  final int? uv;
+  final double? uv;
 
   const DayModel({
     this.maxtempC,
